+++ src/Symfony/Component/Notifier/CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 
  * Make `TransportFactoryTestCase` and `TransportTestCase` compatible with PHPUnit 10+
  * Add `Desktop` channel
+ * Deprecate the `Sms77Transport` class, use `SevenIoTransport`instead
 
 6.3
 ---